Create test showing behavior difference in AGP 7.2 beta04
It seems that previously, lint would not allow reporting issues
on non-source code (ex implicit enum methods), but now it will. This
test shows that a detector that flags all method parameters will now create
issues on the paramter to the valueOf() method, whereas before it would not report
it (can be tested by reverting the build.gradle version bumps).