Add example code to show bug in the Recycle detector
The recycle detector now searchs for animations that never call start(). This detector
does not correctly detect animators created in ternary operators, as shown - in this code,
one of two animators is returned based on a boolean. That animator is then immediately started,
so there is no issue of start() not being called.